requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
loguru>=0.7.0
icecream>=2.1.0
tavern>=2.0.0
python-dotenv>=1.0.0
notion-client>=2.2.0
anthropic>=0.39.0
pyperclip>=1.8.0
httpx>=0.27.0
requests>=2.31.0
rich>=13.0.0
